Removed horizontal lines and calculated mean line
Update: Applied a vertical filter to remove any horizontal lines
Improvement: Instead of showing all the lines, now only the "mean" line
is shown. There is still work to be done to improve this mean estimate.
We probably want to extrapolate this line to the known line segments of
the road